chore: move video into it's own file
