+++ src/itinerary.ts
@@ -1,9 +1,9 @@
 import { Chat } from './chat'
 import { GetItineraryResponse, ItineraryItem, ItineraryPayload } from './event'
-import SubscriptionManager from './subscription-manager'
 import { ConsumerTopic } from './topic'
 import { ExtreamUser } from './user'
 import { InitialResponse, promiseTimeout, TimeStamp } from './utils'
+import { Video } from './video'
 
 export interface RtcConfiguration {
   operators: string[];
@@ -30,36 +30,6 @@ export interface ReadWebRtcResponse {
   payload: ReadWebRtcResponsePayload;
   user: ExtreamUser;
   socketId: string;
-}
-
-export class Video {
-  private socket: SocketIOClient.Socket
-  public id: string
-  public data: ReadWebRtcResponsePayload | null = null
-
-  constructor (socket: SocketIOClient.Socket, id: string) {
-    this.socket = socket
-    this.id = id
-  }
-
-  get (): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-      const callback = (itemData: ReadWebRtcResponse) => {
-        if (itemData.error) {
-          reject(new Error(itemData.error))
-          this.socket.removeEventListener(ConsumerTopic.WebrtcRead, callback)
-        } else if (itemData.payload && itemData.payload.id === this.id) {
-          this.data = itemData.payload
-          resolve()
-          this.socket.removeEventListener(ConsumerTopic.WebrtcRead, callback)
-        }
-      }
-      this.socket.on(ConsumerTopic.WebrtcRead, callback)
-      this.socket.emit(ConsumerTopic.WebrtcRead, {
-        id: this.id
-      })
-    })
-  }
 }
 
 export class Itinerary {
